feat: :construction: Implement retry request
This commit implements a retry for failed request. Due to the serverless nature of HuggingFace model Inference API. The server will only spin up the model on requests. But it will return a HTTP 503 code when the model is still loading. So we will retry the request for some time before a Timeout Exception